rename: statService -> memberStatLogService, productStatService 분리로 인한 클래스 및 필드 네이밍 수정
