+++ be/src/main/java/kr/codesquad/secondhand/api/member/service/MemberProductFacadeService.java
@@ -4,13 +4,12 @@ import java.util.List;
 import java.util.Map;
 import kr.codesquad.secondhand.api.category.domain.Category;
 import kr.codesquad.secondhand.api.category.dto.CategorySummaryResponse;
-import kr.codesquad.secondhand.api.member.dto.request.WishProductRequest;
 import kr.codesquad.secondhand.api.member.dto.response.ProductWishStatusResponse;
 import kr.codesquad.secondhand.api.product.domain.Product;
 import kr.codesquad.secondhand.api.product.domain.ProductStat;
 import kr.codesquad.secondhand.api.product.dto.response.ProductSlicesResponse;
 import kr.codesquad.secondhand.api.product.service.ProductService;
-import kr.codesquad.secondhand.api.product.service.StatService;
+import kr.codesquad.secondhand.api.product.service.ProductStatService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
@@ -24,7 +23,8 @@ import org.springframework.transaction.annotation.Transactional;
 public class MemberProductFacadeService {
 
     private final ProductService productService;
-    private final StatService statService;
+    private final ProductStatService productStatService;
+    private final MemberStatLogRedisService memberStatLogRedisService;
 
 
     @Transactional
@@ -35,7 +35,7 @@ public class MemberProductFacadeService {
 
         List<Product> products = productSlices.getContent();
         Boolean hasNext = productSlices.hasNext();
-        Map<Long, ProductStat> productStats = statService.findProductsStats(products);
+        Map<Long, ProductStat> productStats = productStatService.findProductsStats(products);
 
         return ProductSlicesResponse.of(products, productStats, hasNext);
     }
@@ -44,21 +44,21 @@ public class MemberProductFacadeService {
     public ProductSlicesResponse readMemberWishlist(Long memberId, Long categoryId, Integer page, Integer size) {
         Sort sort = Sort.by(Direction.DESC, "id");
         PageRequest pageRequest = PageRequest.of(page, size, sort);
-        List<Long> productIds = statService.findWishlistByMemberId(memberId);
-        Slice<Product> productSlices = productService.findWishedProductByCategoryIdAndIdIn(productIds, categoryId,
+        List<Long> wishlist = memberStatLogRedisService.findWishlistByMemberId(memberId);
+        Slice<Product> productSlices = productService.findWishedProductByCategoryIdAndIdIn(wishlist, categoryId,
                 pageRequest);
 
         List<Product> products = productSlices.getContent();
         Boolean hasNext = productSlices.hasNext();
-        Map<Long, ProductStat> productStats = statService.findProductsStats(products);
+        Map<Long, ProductStat> productStats = productStatService.findProductsStats(products);
 
         return ProductSlicesResponse.of(products, productStats, hasNext);
     }
 
     @Transactional
     public List<CategorySummaryResponse> readMemberWishCategories(Long memberId) {
-        List<Long> wishProductIds = statService.findWishlistByMemberId(memberId);
-        List<Long> categoryIds = productService.findCategoryIdsByIdIn(wishProductIds);
+        List<Long> wishlist = memberStatLogRedisService.findWishlistByMemberId(memberId);
+        List<Long> categoryIds = productService.findCategoryIdsByIdIn(wishlist);
         List<Category> categories = Category.from(categoryIds);
 
         return CategorySummaryResponse.from(categories);
@@ -66,7 +66,7 @@ public class MemberProductFacadeService {
 
     @Transactional
     public ProductWishStatusResponse checkProductWishedStatus(Long memberId, Long productId) {
-        Boolean isWished = statService.isWishedProductExists(memberId, productId);
+        Boolean isWished = memberStatLogRedisService.containsWishedProductId(memberId, productId);
         return new ProductWishStatusResponse(isWished);
     }
 }
